Fix/polish sidebar related text/functionality
Fixed a missing space in the subtitle under my name in the sidebar, then updated that subtitle in general.

Updated the about section, which includes a link to the contact section, and made links like that use the same animation as the sidebar links. Made non-sidebar links stand out more by making it the same color as the text in that context.

Got rid of the email icon at the bottom of the sidebar, since the subtitle has my email front and center now