Proper fix to about's interaction with Scrollex (see 5eb2a64)
The last fix (5eb2a64) stopped working for no apparent reason, but this led me to actually better understand the problem deeper, and fix it accordingly.

The core of the problem is, the "middle" mode is what I want, but if the about section is too small, it'll never be in the middle (can't scroll up past the top). So, it's area needs to be expanded, and its bottom neighbour contracted. We do this by calculating the viewport's midpoint, and adding/subtracting that accordingly to the areas. This works perfectly.

Only problem is, if the screen's height changes, the padding doesn't update. Making it update would require much more of an edit than I've done here. Thankfully, screens generally don't change in height ever, unless it's a phone going into landscape mode; in such a case, the screen's likely small enough for none of this to matter. and if not, the padding probably only needs to apply to landscape mode. This is such a fringe case that I'm fine if I'm wrong and this semi-minor aesthetic detail breaks a little in said case.