Fix resume replacement bug; rCode is referenced properly
When replacing resumes, the id in the resume url is replaced with the new, desired one. Said ID is targeted by using the default format's rCode, which I've now renamed to "resumeID", since it's explicitly called an ID in the download link (`?export=download&id=...`)

The recent change to formats, which among other things changed the default format's resume ID to "null," meant that resumes would never be replaced since the ID in the url could never be targeted. This fixes that.