Reduce warnings and errors in console (avoid google embeds)
In the process of an earlier commit, I was presented with the idea of warnings and errors causing slower loads. So, just in case it would make the site load faster, in addition to making the console easier to read during development, I set out to prevent as many warnings and as I could.

I soon found that *all* of the errors and warnings came from the youtube embeds and google drive PDF preview of my resume. So I replaced them with alternatives, drastically reducing the number of warnings and errors in the console.

The current setup still throws warnings and errors due to the resume using a google drive preview link. They do not impact the site's functionality whatsoever, but the fact they seem to due to privacy measures blocking XHR requests, tracking cookies, and local storage access makes me think I should perhaps find another way to preview resume PDFs.