Replace remaining gifs with mp4s, style fallbacks
In the process of doing this, I've added the preload attribute to all the video tags. The GIF replacement ones will have that attribute overridden by "autoplay," but I figure it's good to have it there anyway. This will very much help with the longer videos, since they won't be loaded *at all* until clicked.

There are some issues with this setup, since lazy loading is dubious with video tags and such, but I reckon the fact said tags are down to a fifth of the size of what they replace makes up for that in spades.