Implemented code pictures, improved modal telegraphing
Livewire Lifesaver and Experimental VR both have a code snapshot displayed. Said snapshots (and their unused companions) have been renamed for better clarity.

Modal images now have an "expand" icon that animates in when the image is hovered over. The cursor is now "zoom-in" when hovering an unexpanded modal, and "zoom-out" when hovering the area outside the expanded image.

Some pre-existing captions have been updated in the process of switching feature images around.