Disable jekyll on site; fix MIME type error with tailored-format.js
See https://stackoverflow.com/a/59595485 and https://stackoverflow.com/a/76370308; the mime type should at LEAST be `javascript/esm` according to the first link, but the error is saying `text/html`, so I reckon something in the guts of Github Pages is screwing things up. If this commit doesn't fix it, maybe that assumption's incorrect.