All media lazy load/preload none, dynamically set autoplay support
Autoplay makes all lazy loading disregarded, which will become a problem once there are more projects on this site; more media that isn't displayed, but still takes time to load. If autoplay is added *dynamically* once it's shown, you can have your cake and eat it too. ...That is, except at website startup, where you can see the hitch due to javascript needing a moment to load.

A potential solution would be to hide EVERY feature and then show them as if we were reordering the features, regardless if a custom format is requested or not. That's not a today task, whether or not I actually want to go through with it.