Make giflikes properly lazy load always, poster styling
...And by "poster," I mean the background images behind the poster, since you can't actually style the poster itself. Said background images are an optional thumbnail like the normal poster, and then a loading spinner on top of that. This means the gif will be immediately visible if the optional poster is included,  AND look like it's making progress toward being "properly" displayed.

Hiding all features is still an idea, to prevent the display of "wrong" features before the "right" ones are put in place, but I think that's much more of a non-issue and thus not worth "fixing." This wouldn't be immediately visible with that method anyway, which is undesirable, for the moment. Maybe the hitch will be/seem more detrimental some time in the future, just not now.