Added image to taste of culture, different code
In the process, I've learned that PNGs are actually better for "crisp images with a low number of colors," by realizing the jpgs for most of the code screenshots are bigger than the full-size PNGs (less so for the ones with smaller dimensions)! This is because of the compression algorithms they use. Just because JPG is lossy doesn't mean it's always smaller!

I'll have to update all the non "full-size" code snapshots to be PNGs some time. But not now, because it's late, and there are more pressing things to finish.