Add format for Iron Galaxy, fix logic error when reordering features
The classic "wrong way around" kind of if statement error, I had a statement saying "if there are NOT less than or equal to zero desired features" when I intended for that NOT to, well, not be there.

Also minorly adjusted the "role" of A Taste of Culture, to match recent edits to my resume.