Add rockstar format, better align modal hover icons
I didn't think to or didn't want to wrap the img tags just for an icon, but the figures said imgs are contained in vertically center their contents, offsetting the icon (positioned at the top right of the figure, not the img). Took the little extra time to wrap those imgs to fix that.