Update IBM format, refactor format tailoring script
The biggest change here is the introduction of a DebugHelper class, which is just a bit of sugar to reduce the amount of if statements flying around; it contains a method to call console.log() with, but only if it's "enabled."

I've added more comments to the aforementioned script, to make absolutely sure any pair of eyes that may see it in the future can make sense of it.

Now undone, I also made it so the resume preview defaulted to empty by making the URL invalid, with the invalid part replaced in the tailoring process. Doing that WOULD make errors more obvious, and rely less on finding the exact same resume code as the one recorded in the default format, but it'd also leave the site without a fallback when javascript is disabled, and however uncommon that may be, making a proper fallback would be messy and just generally not worth it.