Break override formatting/resolution into its own function
In the process of writing this, I realized that all the fancy string manipulation I do here doesn't matter for anything other than for debug messages, and thus likely isn't worth the computation.

Since it *is* fancy, however, I'm leaving it in this commit, and will only simplify it in the next; that way it's essentially saved for posterity.