Add Double Fine format, fix/adjust resume preview
In a past commit, I accounted for weird extra space in the resume preview on Firefox specifically. This no longer happens, but the preview still looked strange; it's flush against the top edge, but has gaps around the rest of it.
The gaps wouldn't be so weird if there was one on the top, but there isn't, on both my Firefox and Chrome. So, I've done some relative positioning to hide them.